combine locator and resolver class
