+++ tests/unit/Files/PathResolverTest.php
@@ -22,4 +22,20 @@ class PathResolverTest extends PHPUnit_Framework_TestCase
 		$expected = ['/other_root/bar', '/root/foo/bar'];
 		$this->assertEquals($expected, $resolver->resolve('foo/bar'));
 	}
+
+	/** @test */
+	public function locatesFiles()
+	{
+		$basenames = [
+			TESTS_RSC_DIR.'/files/root/filename',
+			TESTS_RSC_DIR.'/files/other_root/filename',
+		];
+		$extensions = ['.foo', '.bar'];
+		$paths = $this->makeResolver()->locate($basenames, $extensions);
+		$expected = [
+			TESTS_RSC_DIR.'/files/root/filename.foo',
+			TESTS_RSC_DIR.'/files/other_root/filename.bar',
+		];
+		$this->assertEquals($expected, $paths);
+	}
 }
